Reduced level of detail in header file docblocks.
The header file used to document all permitted return values for conversion
functions. This level of detail will inhibit future development of the library
if clients come to depend on this behaviour. I've reduced the level of detail
simply to itemise the return values and their causes, and added a warning not
to trust the byte stream if a conversion fails.