Modified bs_malloc_output to null-terminate buffer.
bs_malloc_output is used to allocate memory for outputting data. This data is
most often string data which needs to be null terminated, and it's a pain for
individual save functions to have to:
 - remember to add the null
 - mess around with the length parameter to return the right thing

This commit changes the semantics of bs_malloc_output(). It now allocates a
buffer one byte larger than required, and writes a null terminator into that
extra byte.

String output works easily - allocate the number of bytes needed and the null
will silently be added with no extra work. The correct string length
is returned.

Binary output isn't affected either - adding an extra null doesn't make any
difference as the correct number of bytes is returned.