Started implementation of a streaming system
I'd like the stream to be able to buffer incoming data and process it in
sections. The name "byte stream" implies that some sort of streaming ought to
be possible, so it's about time it started...

This commit adds to struct BS to maintain the stream state internally, and adds
tests to make sure this is set to zero when allocating.