Changed bs_stream_empty to bs_stream_reset, and stopped it returning data
bs_stream_empty() returns any unstreamed bytes, but I'm not sure this is a
worthwhile use case to support - surely unstreamed bytes should be processed
tidily via a call to bs_flush()?

I've removed this functionality, and changed the function name to more
accurately express what's going on. This can always be replaced if necessary (I
think it's likely it'll be necessary to provide access to the stream count for
some purpose).