Renamed bytestream processing functions
Renamed bs_walk -> bs_map, as well as all supplied walk functions.
Renamed bs_accumulate -> bs_fold, as well as all supplied accumulate functions.

This change to nomenclature is to match the terminology to that used commonly
in functional-programming land (specifically with Haskell). A conversion from
one bytestream to another _is_ a mapping, and fold is a clearer description of
what actually happens.